LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/try-tactic-margin-expected.txt


PASS --pf
PASS --pf flip-block
PASS --pf flip-inline
PASS --pf flip-block flip-inline
FAIL --pf flip-start assert_equals: offsetLeft expected 25 but got 45
FAIL --pf flip-block flip-start assert_equals: offsetLeft expected 335 but got 45
FAIL --pf flip-inline flip-start assert_equals: offsetLeft expected 25 but got 325
FAIL --pf flip-block flip-inline flip-start assert_equals: offsetLeft expected 335 but got 325
